valid code

+++ WebShopREST/src/services/MembershipService.java
@@ -147,7 +147,7 @@ public class MembershipService {
 			if(existingPromoCode == null) {
 				throw new WebApplicationException(Response.status(Status.CONFLICT).entity("Promo kod ne vazi").build());
 			}
-			if(existingPromoCode.getUsageCount() > 0) {
+			if(existingPromoCode.getUsageCount() > 0 && existingPromoCode.getValidDate().compareTo(new Date()) <= 0) {
 				discount = existingPromoCode.getDiscountPercentage() / 100.0;			
 				existingPromoCode.reduceUsgeCount();
 				promoCodeDao.writeFile();
